feat: разнесена логика моделей и взаимодействия с ними
models.py: все функции вынесены в models_utils.py.
Добавлен models_utils.py.
