+++ models.py
@@ -2,7 +2,6 @@ from typing import Type
 
 from tortoise.models import Model
 from tortoise import fields
-from tortoise.exceptions import DoesNotExist
 
 
 class User(Model):
@@ -12,15 +11,6 @@ class User(Model):
 
     def __str__(self) -> str:
         return f'Bot user (id: {self.id}, discord_id: {self.discord_id})'
-
-
-async def get_user_model_by_discord_id(discord_id: int) -> User:
-
-    try:
-        return await User.get(discord_id=discord_id)
-
-    except DoesNotExist:
-        return await User.create(discord_id=discord_id)
 
 
 class QuestionBase(Model):
@@ -102,27 +92,6 @@ class QuestionAnother(QuestionBase):
 
     class Meta:
         table = 'question_another'
-
-
-async def get_question_by_discord_channel_id(discord_channel_id: int) -> Type[QuestionBase] | None:
-    for question_model in (QuestionThemeLesson, QuestionProject, QuestionAnother):
-
-        try:
-            return await question_model.get(discord_channel_id=discord_channel_id)
-
-        except DoesNotExist:
-            continue
-
-
-async def get_all_questions(**filter_kwargs) -> list[QuestionThemeLesson | QuestionProject | QuestionAnother]:
-
-    result = []
-
-    result.extend(await QuestionThemeLesson.filter(**filter_kwargs))
-    result.extend(await QuestionProject.filter(**filter_kwargs))
-    result.extend(await QuestionAnother.filter(**filter_kwargs))
-
-    return result
 
 
 class QuestionStatistics(Model):
